+++ src/main/java/com/example/bugbug/controller/UserController.java
@@ -5,21 +5,19 @@ import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
+import com.example.bugbug.form.LoginForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.InitBinder;
+import org.springframework.web.bind.annotation.*;
 import com.example.bugbug.entity.Users;
 import com.example.bugbug.form.InputForm;
 import com.example.bugbug.service.UserService;
 import com.example.bugbug.validator.MailValidator;
 import com.example.bugbug.validator.PassValidator;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import lombok.RequiredArgsConstructor;
@@ -30,6 +28,7 @@ public class UserController {
 	// インスタンス作成部分
 	private final UserService service;
 	private final PassValidator passValidator;
+	private final PassValidator mailValidator;
 
 	@Autowired
 	HttpSession session;
@@ -69,7 +68,7 @@ public class UserController {
 	public String createAcount(@Validated InputForm f, BindingResult bindingResult, Model model) {
 		// validation
 		if (bindingResult.hasErrors()) {
-			return "sinup";
+			return "sinup/form";
 		}
 		// パスワードのハッシュ化
 		String hash = service.hash(f.getPass());
@@ -85,31 +84,5 @@ public class UserController {
 		session.setAttribute("user_id", list.get(0).getUser_id());
 		session.setAttribute("user_name", f.getName());
 		return "redirect:/index";
-	}
-
-	// ログイン画面遷移
-	@RequestMapping(value = "login/form")
-	public String viewLoginForm() {
-		return "login";
-	}
-
-	// ログイン処理
-	@PostMapping(value = "login")
-	public String authLogin(@Validated LoginForm f, BindingResult bindingResult, Model model,
-			RedirectAttributes redirectAttributes) {
-
-		// メールで検索
-		List<Users> list = service.findMail(f.getMail());
-		// メールとパスワードが正しいとき
-		if (!list.isEmpty() && service.match(list.get(0).getPass(), f.getPass())) {
-			// sessionに値を登録
-			session.setAttribute("user_id", list.get(0).getUser_id());
-			session.setAttribute("user_name", list.get(0).getName());
-			model.addAttribute("msg", "ログイン成功");
-			return "redirect:/index";
-		} else {
-			redirectAttributes.addFlashAttribute("msg", "メールアドレスかパスワードが違います");
-			return "redirect:/login/form";
-		}
 	}
 }
