+++ src/main/java/com/example/bugbug/service/AuthService.java
@@ -1,0 +1,4 @@
+package com.example.bugbug.service;
+
+public interface AuthService {
+}
