+++ src/main/java/com/example/bugbug/service/AuthServiceImpl.java
@@ -1,0 +1,4 @@
+package com.example.bugbug.service;
+
+public class AuthServiceImpl {
+}
